React/Nest-Pro/CHAT-App/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
aiofiles==23.2.1
google-generativeai==0.3.2
httpx==0.25.2
